requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
filelock==3.8.2
huggingface-hub==0.11.1
idna==3.4
joblib==1.2.0
packaging==22.0
PyYAML==6.0
requests==2.28.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
